fix: Fix loop gap for MP3 files from SWF
feat: introduce right resampling and gap-correction for AudioBuffers

Used metadata from SWF for allow looping mp3 files without gap.

Followed mp3 spec, all encoders add silent region in the beginning and ending of mpeg files.
This not so easy to determine, because `silent` is not really silent, because it store trashed not zero data.